+++ CHANGELOG.md
@@ -222,6 +222,7 @@ Each release is divided into the following main categories:
 
 ### FIXES
 
+- Swiper Fullscreen: Hitting the ESC key closed a form modal. Fixed!
 - Btn AJAX Loader: Now executed with a delay, as it caused button size issues when the page wasn't fully rendered yet.
 - Admin Modal: Image preview now bigger.
 - Softpage (Firefox bug): Any selected text is now "deselected" when openen the softpage.